second

+++ pfa5/health/health-app/src/main/java/ma/hopital/healthapp/dao/PatientDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ma.hopital.healthapp.model.Patient;
 
-public interface PatientDao extends JpaRepository<Patient, Long> {
-	Patient findByPatientId(Long id);
+public interface PatientDao extends JpaRepository<Patient, Integer> {
+	Patient findByPatientId(int id);
 }
