+++ pfa5/health/health-app/src/main/java/ma/hopital/healthapp/dao/SpecialiteDao.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ma.hopital.healthapp.model.Specialite;
 
-public interface SpecialiteDao extends JpaRepository<Specialite, Long>{
-	Specialite findBySpecialiteId(Long id);
+public interface SpecialiteDao extends JpaRepository<Specialite, Integer>{
+	Specialite findBySpecialiteId(int id);
 	Specialite findByLibelle(String libelle);
 }
